+++ src/auth-state-router/auth-state-router.service.js
@@ -70,6 +70,7 @@
             } else if(toState.accessRights && !authorizationService.hasRights(toState.accessRights, toState.areAllRightsRequired)) {
                 // checking rights to enter state
                 event.preventDefault();
+                loadingModalService.close();
                 alertService.error('openlmisAuth.authorization.error');
             }
         }